+++ packages/vue-simulator-renderer/src/utils/get-client-rects.ts
@@ -1,9 +1,11 @@
-import { isElement } from '@alilc/lowcode-utils';
+import { isElement } from '@knxcloud/lowcode-utils';
 
 // a range for test TextNode clientRect
 const cycleRange = document.createRange();
 
 export function getClientRects(node: Element | Text) {
+  if (!node.parentNode) return [];
+
   if (isElement(node)) {
     return [node.getBoundingClientRect()];
   }